feat: StockcorrectionBody component
